Added production validation for person_professions and professions controllers

+++ app/controllers/person_professions.py
@@ -2,11 +2,17 @@ from flask import request, jsonify
 from app import application as app
 from app.models.person_professions import *
 from app.const import *
+from app import isOnDev
 
 @app.route('/person-profession', methods=['GET', 'POST'])
 def person_profession_fun():
 
-    construct = {}
+    construct = {
+        'success': False,
+        'message': 'Method not allowed :)'
+    }
+    response = jsonify(construct)
+    response.status_code = HttpStatus.NOT_ALLOWED
 
     #   Get all from table person_professions
     if request.method == 'GET':
@@ -18,7 +24,7 @@ def person_profession_fun():
         response.status_code = HttpStatus.OK
 
     #   Trying to insert into the table
-    elif request.method == 'POST':
+    elif request.method == 'POST' and isOnDev:
 
         #   Trying to get parameters from the POST method
         try:
@@ -61,6 +67,13 @@ def person_profession_fun():
 @app.route('/person-profession/<int:person_profession_id>', methods=['GET', 'PUT', 'DELETE'])
 def personProfessionID(person_profession_id):
 
+    construct = {
+        'success': False,
+        'message': 'Method not allowed :)'
+    }
+    response = jsonify(construct)
+    response.status_code = HttpStatus.NOT_ALLOWED
+
     #   Trying to get the area with person_profession_id
     person_profession = Person_Profession.query.filter_by(person_profession_id=person_profession_id).first()
 
@@ -89,8 +102,7 @@ def personProfessionID(person_profession_id):
         response.status_code = HttpStatus.OK
 
     #   Update their data
-    elif request.method == 'PUT':
-        construct = {}
+    elif request.method == 'PUT' and isOnDev:
 
         #   Trying to get parameters from the PUT method
         try:
@@ -130,8 +142,8 @@ def personProfessionID(person_profession_id):
             response.status_code = HttpStatus.BAD_REQUEST
 
     #   Delete it from the database
-    elif request.method == 'DELETE':
-        construct = {}
+    elif request.method == 'DELETE' and isOnDev:
+
         try:
             person_profession.delete()
             construct['success'] = True
@@ -143,4 +155,5 @@ def personProfessionID(person_profession_id):
             construct['error'] = str(e)
             response = jsonify(construct)
             response.status_code = HttpStatus.BAD_REQUEST
+
     return response
